+++ packages/parameterized_test/lib/src/test_options/test_options_ext.dart
@@ -1,13 +1,14 @@
+//ignore_for_file: long-parameter-list
 import 'package:test/test.dart' as dart_test;
 
 import 'test_options.dart';
 import 'value_with_test_options.dart';
 
-extension IterableTestParametersEx<T extends Iterable> on T {
+extension IterableTestParametersEx<T extends Iterable<dynamic>> on T {
   ValueWithTestOptions withTestOptions({
     dart_test.Timeout? timeout,
-    skip,
-    tags,
+    dynamic skip,
+    dynamic tags,
     Map<String, dynamic>? onPlatform,
     int? retry,
   }) =>
@@ -26,8 +27,8 @@ extension IterableTestParametersEx<T extends Iterable> on T {
 extension TestParametersEx<T extends Object> on T {
   ValueWithTestOptions withTestOptions({
     dart_test.Timeout? timeout,
-    skip,
-    tags,
+    dynamic skip,
+    dynamic tags,
     Map<String, dynamic>? onPlatform,
     int? retry,
   }) =>