+++ packages/parameterized_test/lib/src/util/test_value_helpers.dart
@@ -1,7 +1,7 @@
 import '../errors/parameter_count_error.dart';
 import '../test_options/value_with_test_options.dart';
 
-void mapTests<T>(Iterable<ValueWithTestOptions<T>> values, int length, Function(Iterable) body) {
+void mapTests(Iterable<ValueWithTestOptions> values, int length, dynamic Function(Iterable<dynamic>) body) {
   for (final ValueWithTestOptions value in values) {
     value.testOptions.test('$value', () {
       validityCheck(value, length);
@@ -10,7 +10,7 @@ void mapTests<T>(Iterable<ValueWithTestOptions<T>> values, int length, Function(
   }
 }
 
-void validityCheck<T>(Iterable<T> values, int length) {
+void validityCheck(Iterable<dynamic> values, int length) {
   if (values.length != length) {
     throw ParameterCountError(length, values.length);
   }